Add Windows support for most operations.
AMDeviceMountImage doesn't seem to be available in MobileDevice.dll on
Windows.  That is the only feature that should be missing.

I believe that all of the necessary dependencies are installed as part
of iTunes, but it's possible that the iPhone Configuration Utility is
necessary.  I think a 32-bit python.exe is currently required.