Updated version to 1.9.3 (from 2.x).
master branch was intended to be the latest and greatest code for the
2.x re-factor (commands into their own files), but that has been
shelved. 1.x was essentially code from master cherry-picked, and this
has proven problematic in keeping in sync, so I am simplifying things.

master branch was essentially 1.x anyway since there was no active
development on 2.x according to the former roadmap.