Skip removing FIFOs from device (#528)
* Determine if a remote file is a FIFO

* Skip removing FIFOs from device

That operation always fails with error `0xc` anyway.

To distinguish regular files from FIFO, new `READ_DIR_FIFO` is added to
`read_dir_cb_reason`. The behavior of other operations regarding FIFOs
is left the same by supporting the new reason.

Another implementation could be not to extend the `read_dir_cb_reason`
and add another flag to `read_dir`'s `callback`, but that would allow
meaningless cases, such as `reason = READ_DIR_AFTER_DIR, is_fifo =
true`.

The ideal implementation would be to extent the `READ_DIR_FILE` enum
case with a flag (i.e., `READ_DIR_FILE(is_fifo = true)`), but that is
not supported in Objective-C.

* Continue removing files after an error

The fix is to skip removing FIFOs had a side-effect of failing to remove
a directory where a skipped file is because the directory is not empty.

The change here is to continue removing files after errors, which is
also helpful in cases when `ios-deploy` fails to remove `/Documents/`
("Error 0xa: You do not have permission."), but still needs to continue
removing what it can from `/Library/`.

* Fix printing error messages

This makes `ios-deploy` print error messages for known error codes so
that the console log looks like:

```
ios-deploy[13450:4855044] [ !! ] Error 0x1: An unknown error occurred. AFCRemovePath(conn, name)
ios-deploy[13450:4855044] [ !! ] Error 0xa: You do not have permission. AFCRemovePath(conn, name)
```

instead of:

```
ios-deploy[13409:4853438] [ !! ] Error 0x1: unknown. AFCRemovePath(conn, name)
ios-deploy[13409:4853438] [ !! ] Error 0xa: unknown. AFCRemovePath(conn, name)
```

The fix is to create an error code from the input `error` to match
what's defined in `errorcode_to_id` (for some reason, the errors are
`OR`'d with `0xe8000000`). This change is taken from the
`SDMMobileDevice` project, which is the source for the error codes.

Co-authored-by: Gabe Giosia <giosia@google.com>